+++ common/components/Articles/ArticlesList.tsx
@@ -1,0 +1,16 @@
+import { ArticlesItem } from "./ArticlesItem";
+
+type Props = {
+  articles: Article[];
+};
+export const ArticlesList = ({ articles }: Props) => {
+  return (
+    <>
+      {articles.map((article, index) => (
+        <div key={index} className="mb-8">
+          <ArticlesItem article={article} key={index} />
+        </div>
+      ))}
+    </>
+  );
+};