+++ common/components/Icons/ChevronRight.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+export function ChevronRight({ ...props }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="currentColor"
+      className="bi bi-chevron-right"
+      viewBox="0 0 12 16"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        d="M4.646 1.646a.5.5 0 01.708 0l6 6a.5.5 0 010 .708l-6 6a.5.5 0 01-.708-.708L10.293 8 4.646 2.354a.5.5 0 010-.708"
+      ></path>
+    </svg>
+  );
+}