+++ common/components/layout/PageHead.tsx
@@ -1,0 +1,66 @@
+import React, { useEffect } from "react";
+import Head from "next/head";
+
+function isElementNearViewport(element: any) {
+  const rect = element.getBoundingClientRect();
+  if (parseInt(rect.bottom) <= window.innerHeight + 600) {
+    return true;
+  }
+}
+
+export const PageHead = ({
+  article,
+  title,
+  description,
+}: {
+  article?: Article;
+  title: string;
+  description?: string;
+}) => {
+  useEffect(() => {
+    (window as any).lazyLoadElements = [];
+
+    window.addEventListener("scroll", () => {
+      (window as any).lazyLoadElements.forEach(
+        (element: HTMLIFrameElement, index: number) => {
+          if (isElementNearViewport(element)) {
+            element.src = element.dataset.src ?? "";
+            // Remove this element from the list
+            (window as any).lazyLoadElements.splice(index, 1);
+          }
+        }
+      );
+    });
+    document.querySelectorAll("iframe[data-src]").forEach((element, index) => {
+      (window as any).lazyLoadElements[index] = element;
+      //   const observer = new IntersectionObserver(function (entries) {
+      //     if (entries[0].isIntersecting) {
+      //       console.log("Elvis has ENTERED the building ");
+      //       (element as HTMLIFrameElement).src =
+      //         (element as HTMLIFrameElement).dataset.src ?? "";
+      //       observer.unobserve(element);
+      //     }
+      //   });
+
+      //   observer.observe(element);
+    });
+  }, []);
+  return (
+    <Head>
+      <title>{title}</title>
+      {article ? (
+        <>
+          <meta name="description" content={article.meta.description} />
+          <meta
+            property="article:published_time"
+            content={article.publish_date}
+          />
+        </>
+      ) : (
+        <meta name="description" content={description} />
+      )}
+
+      <link rel="icon" href="/favicon.ico" />
+    </Head>
+  );
+};