+++ data/articles/markdown/index.ts
@@ -1,0 +1,140 @@
+import you_dont_need_to_meditate from "./you-dont-need-to-meditate.md";
+import it_starts_with_you from "./it-starts-with-you.md";
+import there_is_only_now from "./there-is-only-now.md";
+import the_secret_of_zen from "./the-secret-of-zen.md";
+import the_purpose_of_meditation from "./the-purpose-of-meditation.md";
+import the_attachment_paradox from "./the-attachment-paradox.md";
+import overcoming_worry from "./overcoming-worry.md";
+import dont_force_life from "./dont-force-life.md";
+import you_cant_completely_control_your_mind from "./you-cant-completely-control-your-mind.md";
+import you_cant_completely_control_suffering from "./you-cant-completely-control-suffering.md";
+import emotional_attachment_causes_psychological_suffering from "./emotional-attachment-causes-psychological-suffering.md";
+const articles: Article[] = [
+  {
+    title: "You Don’t Need To Meditate",
+    slug: "you-dont-need-to-meditate",
+    markdown: you_dont_need_to_meditate,
+    excerpt:
+      "It’s far more important to apply the fundamentals from Buddhism to your life than to practice meditation; **one moment of understanding could be more valuable than a lifetime of daily meditation if...",
+    categories: ["meditation"],
+    meta: {
+      title: "You Don’t Need To Meditate",
+      description:
+        "It’s far more important to apply the fundamentals from Buddhism to your life than to practice meditation; **one moment of understanding could be more valuable than a lifetime of daily meditation if...",
+    },
+    images: {
+      cover: {
+        src: "mountain_temple.jpg",
+        alt: "Peaceful Mountain Temple",
+        color: "#4F3A63",
+      },
+    },
+    publish_date: "2024-07-10",
+    next_article: {
+      title: "It Starts With You",
+      slug: "it-starts-with-you",
+      markdown: "__MARKDOWN_IMPORT__",
+      excerpt:
+        "Studying modern Buddhism is hard work. There are so many books on the topic, it’s actually very difficult to get to the core message of Buddhism. Even Reddit’s r/zen is a reportedly...",
+      categories: ["meditation"],
+      meta: {
+        title: "It Starts With You",
+        description:
+          "Studying modern Buddhism is hard work. There are so many books on the topic, it’s actually very difficult to get to the core message of Buddhism. Even Reddit’s r/zen is a reportedly...",
+      },
+      images: {
+        cover: {
+          src: "mountains_tapestry.jpg",
+          alt: "Peaceful Mountain Temple",
+          color: "#4F3A63",
+        },
+      },
+      publish_date: "2024-08-10",
+    },
+  },
+  {
+    title: "It Starts With You",
+    slug: "it-starts-with-you",
+    markdown: it_starts_with_you,
+    excerpt:
+      "Studying modern Buddhism is hard work. There are so many books on the topic, it’s actually very difficult to get to the core message of Buddhism. Even Reddit’s r/zen is a reportedly...",
+    categories: ["meditation"],
+    meta: {
+      title: "It Starts With You",
+      description:
+        "Studying modern Buddhism is hard work. There are so many books on the topic, it’s actually very difficult to get to the core message of Buddhism. Even Reddit’s r/zen is a reportedly...",
+    },
+    images: {
+      cover: {
+        src: "mountains_tapestry.jpg",
+        alt: "Peaceful Mountain Temple",
+        color: "#4F3A63",
+      },
+    },
+    publish_date: "2024-08-10",
+    next_article: {
+      title: "There is Only Now",
+      slug: "there-is-only-now",
+      markdown: "__MARKDOWN_IMPORT__",
+      excerpt:
+        "Understanding your relationship with time is one of the most important and liberating concepts you can explore. There’s a lot that we take for granted about time which, when understood, can instantly...",
+      categories: ["meditation"],
+      meta: {
+        title: "There is Only Now",
+        description:
+          "Understanding your relationship with time is one of the most important and liberating concepts you can explore. There’s a lot that we take for granted about time which, when understood, can instantly...",
+      },
+      images: {
+        cover: {
+          src: "atmospheric_clock.jpg",
+          alt: "Peaceful Mountain Temple",
+          color: "#4F3A63",
+        },
+      },
+      publish_date: "2024-09-10",
+    },
+  },
+  {
+    title: "There is Only Now",
+    slug: "there-is-only-now",
+    markdown: there_is_only_now,
+    excerpt:
+      "Understanding your relationship with time is one of the most important and liberating concepts you can explore. There’s a lot that we take for granted about time which, when understood, can instantly...",
+    categories: ["meditation"],
+    meta: {
+      title: "There is Only Now",
+      description:
+        "Understanding your relationship with time is one of the most important and liberating concepts you can explore. There’s a lot that we take for granted about time which, when understood, can instantly...",
+    },
+    images: {
+      cover: {
+        src: "atmospheric_clock.jpg",
+        alt: "Peaceful Mountain Temple",
+        color: "#4F3A63",
+      },
+    },
+    publish_date: "2024-09-10",
+    next_article: {
+      title: "The Secret of Zen",
+      slug: "the-secret-of-zen",
+      markdown: "__MARKDOWN_IMPORT__",
+      excerpt:
+        "Zen by design is provocative. It’s intended this way to invoke a reaction or realisation from you; a moment of spontaneous enlightenment....",
+      categories: ["meditation"],
+      meta: {
+        title: "The Secret of Zen",
+        description:
+          "Zen by design is provocative. It’s intended this way to invoke a reaction or realisation from you; a moment of spontaneous enlightenment....",
+      },
+      images: {
+        cover: {
+          src: "mountain_temple.jpg",
+          alt: "Peaceful Mountain Temple",
+          color: "#4F3A63",
+        },
+      },
+      publish_date: "2024-10-10",
+    },
+  },
+];
+export default articles;