+++ next.config.js
@@ -2,10 +2,37 @@
 const nextConfig = {
   reactStrictMode: false,
   swcMinify: true,
-  experimental: {
-    images: {
-      unoptimized: true,
-    },
+  webpack: (config) => {
+    return Object.assign({}, config, {
+      module: Object.assign({}, config.module, {
+        rules: config.module.rules.concat([
+          {
+            test: /\.md$/,
+            loader: "raw-loader",
+          },
+        ]),
+      }),
+    });
+  },
+  images: {
+    loader: "custom",
+    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
+    deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
+  },
+  env: {
+    nextImageExportOptimizer_imageFolderPath: "public/images/global",
+    nextImageExportOptimizer_exportFolderPath: "out",
+    nextImageExportOptimizer_quality: 90,
+    nextImageExportOptimizer_storePicturesInWEBP: true,
+
+    // If you do not want to use blurry placeholder images, then you can set
+    // nextImageExportOptimizer_generateAndUseBlurImages to false and pass
+    // `placeholder="empty"` to all <ExportedImage> components.
+    //
+    // If nextImageExportOptimizer_generateAndUseBlurImages is false and you
+    // forget to set `placeholder="empty"`, you'll see 404 errors for the missing
+    // placeholder images in the console.
+    nextImageExportOptimizer_generateAndUseBlurImages: true,
   },
 };