+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 0.5.0
+
+- First article reviewed and updated, first draft
+
+# 0.4.2
+
+- Fix config for optimised images, goodbye yarn
+
+# 0.4.1
+
+- Use sharp@0.31.1
+
+# 0.4.0
+
+- Attempt to fix optimised images
+
+# 0.3.0
+
+- First blog article, blog index and blog view
+
 # 0.2.0
 
 - "Coming soon..." holding page